Allows to modify file depth from an external caller
We use Log4Go in our project with a centralized wrapper that, among other things executes `log4go`'s logging. If `DefaultFileDepth` is a constant, we will see the name of the package as the wrapper one, instead of the original caller, which is one level up. With this modification we can modify `DefaultFileDepth` outside, so it passes a `3`, that'll allows us to track an additional level up and not the wrapper package.